# internal/features/tasks/transport/http/create_task.go
package tasks_transport_http

import (
	"net/http"

	"github.com/eugeniuszglinski/golang-todoapp/internal/core/domain"
	core_logger "github.com/eugeniuszglinski/golang-todoapp/internal/core/logger"
	core_http_request "github.com/eugeniuszglinski/golang-todoapp/internal/core/transport/http/request"
	core_http_response "github.com/eugeniuszglinski/golang-todoapp/internal/core/transport/http/response"
)

type CreateTaskRequest struct {
	Title        string  `json:"title" validate:"required,min=1,max=100"          example:"Homework"`
	Description  *string `json:"description" validate:"omitempty,min=1,max=1000"  example:"Do it before Friday"`
	AuthorUserID int     `json:"author_user_id" validate:"required"               example:"1"`
}

type CreateTaskResponse TaskDtoResponse

// CreateTask   godoc
// @Summary     Create a new task
// @Description Create a new task with the provided details
// @Tags        tasks
// @Accept      application/json
// @Produce     application/json
// @Param       request body CreateTaskRequest true "CreateTask request body"
// @Success     201 {object} CreateTaskResponse "Created user"
// @Failure     400 {object} core_http_response.ErrorResponse "Bad request"
// @Failure     404 {object} core_http_response.ErrorResponse "Author not found"
// @Failure     500 {object} core_http_response.ErrorResponse "Internal server error"
// @Router      /tasks [post]
func (h *TasksHttpHandler) CreateTask(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	logger := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHttpResponseHandler(logger, w)

	var request CreateTaskRequest
	if err := core_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(err, "failed to decode and validate request")

		return
	}

	taskDomain := domain.NewTaskUninitialized(request.Title, request.Description, request.AuthorUserID)

	createdTaskDomain, err := h.tasksService.CreateTask(ctx, taskDomain)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to create task")

		return
	}

	response := CreateTaskResponse(*taskDtoFromDomain(createdTaskDomain))

	responseHandler.JsonResponse(response, http.StatusCreated)
}